Ignoring reset requests from spectators (and non-players). Storing current date and server's performance in the database. Changed datetimes of games to timestamps in the database. Changed database schema to keep only max values (economy and infrastructure). Added company identifier to the list of clients (via Slack).

+++ genowefa/src/main/java/de/miltschek/genowefa/CustomServerListener.java
@@ -137,7 +137,7 @@ public class CustomServerListener extends ServerListenerAdapter {
 			int days = date.getRawValue() - this.lastDate.getRawValue();
 			if (days > 0) {
 				long realTime = System.currentTimeMillis() - lastDateTs;
-				this.context.setPerformance(realTime / days);
+				this.context.setPerformance((int)(realTime / days));
 			}
 		}
